requirements.txt
# SmartLocker Cloud Backend
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
aiosqlite==0.20.0
alembic==1.13.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
python-multipart==0.0.9
python-dotenv==1.0.1
jinja2==3.1.4
aiofiles==24.1.0
httpx==0.27.0
pydantic[email]==2.9.0
pydantic-settings==2.5.0
